+++ src/main/java/org/deepsymmetry/beatlink/dbserver/BinaryField.java
@@ -20,12 +20,6 @@ public class BinaryField extends Field {
     private final byte typeTag = (byte)0x14;
 
     /**
-     * The byte which identifies this type of field within a message's argument list, which for some reason is
-     * different than the type tag itself.
-     */
-    private final byte argumentTag = (byte)0x03;
-
-    /**
      * The number of bytes making up the network representation of the value of this field, excluding the type tag.
      */
     private final int size;
@@ -113,7 +107,7 @@ public class BinaryField extends Field {
 
     @Override
     public byte getArgumentTag() {
-        return argumentTag;
+        return (byte) 0x03;
     }
 
     @Override